Move slave clock config into slave_lib
