+++ janus/main.c
@@ -5,8 +5,7 @@
 #include "fnr.h"
 #include "gpio.h"
 #include "sevenSeg.h"
-
-void SystemClock_Config(void);
+#include "systemClock.h"
 
 int main(void)
 {
@@ -38,49 +37,6 @@ int main(void)
         }
     }
 }
-
-/**
-* @brief  System Clock Configuration
-*         The system Clock is configured as follow : 
-*            System Clock source            = PLL (HSI)
-*            SYSCLK(Hz)                     = 48000000
-*            HCLK(Hz)                       = 48000000
-*            AHB Prescaler                  = 1
-*            APB1 Prescaler                 = 1
-*            HSI Frequency(Hz)              = 8000000
-*            PREDIV                         = 1
-*            PLLMUL                         = 6
-*            Flash Latency(WS)              = 1
-* @param  None
-* @retval None */
-void SystemClock_Config(void)
-{
-    RCC_ClkInitTypeDef RCC_ClkInitStruct;
-    RCC_OscInitTypeDef RCC_OscInitStruct;
-
-    /* No HSE Oscillator on Nucleo, Activate PLL with HSI as source */
-    RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_NONE;
-    RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
-    RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSI;
-    RCC_OscInitStruct.PLL.PREDIV = RCC_PREDIV_DIV1;
-    RCC_OscInitStruct.PLL.PLLMUL = RCC_PLL_MUL6;
-    if (HAL_RCC_OscConfig(&RCC_OscInitStruct)!= HAL_OK)
-    { /* Initialization Error */
-        while(1); 
-    }
-
-    /* Select PLL as system clock source and configure the HCLK, PCLK1 clocks dividers */
-    RCC_ClkInitStruct.ClockType = (RCC_CLOCKTYPE_SYSCLK | RCC_CLOCKTYPE_HCLK | RCC_CLOCKTYPE_PCLK1);
-    RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_PLLCLK;
-    RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV1;
-    RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV1;
-    if (HAL_RCC_ClockConfig(&RCC_ClkInitStruct, FLASH_LATENCY_1)!= HAL_OK)
-    {
-        /* Initialization Error */
-        while(1); 
-    }
-}
-
 
 #ifdef  USE_FULL_ASSERT
