+++ slave_lib/peripherals/systemClock.h
@@ -1,0 +1,6 @@
+#ifndef SYSTEMCLOCK_H_
+#define SYSTEMCLOCK_H_
+
+void SystemClock_Config(void);
+
+#endif
